requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
face-recognition==1.3.0
pillow==10.1.0
python-jose==3.3.0
pydantic==2.5.2 